+++ src/test/java/org/stellar/sdk/TransactionTest.java
@@ -73,7 +73,7 @@ public class TransactionTest {
             },
             null,
             new TransactionPreconditions(
-                null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>(), null),
+                null, null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>()),
             null,
             Network.PUBLIC);
 
@@ -118,7 +118,7 @@ public class TransactionTest {
             },
             null,
             new TransactionPreconditions(
-                null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>(), null),
+                null, null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>()),
             null,
             Network.PUBLIC);
 
@@ -158,7 +158,7 @@ public class TransactionTest {
             },
             null,
             new TransactionPreconditions(
-                null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>(), null),
+                null, null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>()),
             null,
             Network.PUBLIC);
 
@@ -197,7 +197,7 @@ public class TransactionTest {
             },
             null,
             new TransactionPreconditions(
-                null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>(), null),
+                null, null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>()),
             null,
             Network.TESTNET);
 
@@ -278,7 +278,7 @@ public class TransactionTest {
             new org.stellar.sdk.operations.Operation[] {invokeHostFunctionOperation},
             null,
             new TransactionPreconditions(
-                null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>(), null),
+                null, null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>()),
             sorobanData,
             Network.TESTNET);
 
@@ -314,7 +314,7 @@ public class TransactionTest {
             new org.stellar.sdk.operations.Operation[] {operation},
             null,
             new TransactionPreconditions(
-                null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>(), null),
+                null, null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>()),
             null,
             Network.TESTNET);
     assertTrue(transaction.isSorobanTransaction());
@@ -351,7 +351,7 @@ public class TransactionTest {
             new org.stellar.sdk.operations.Operation[] {operation},
             null,
             new TransactionPreconditions(
-                null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>(), null),
+                null, null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>()),
             sorobanData,
             Network.TESTNET);
     assertTrue(transaction.isSorobanTransaction());
@@ -387,7 +387,7 @@ public class TransactionTest {
             new org.stellar.sdk.operations.Operation[] {operation},
             null,
             new TransactionPreconditions(
-                null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>(), null),
+                null, null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>()),
             sorobanData,
             Network.TESTNET);
     assertTrue(transaction.isSorobanTransaction());
@@ -416,7 +416,7 @@ public class TransactionTest {
             new org.stellar.sdk.operations.Operation[] {operation, operation},
             null,
             new TransactionPreconditions(
-                null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>(), null),
+                null, null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>()),
             null,
             Network.TESTNET);
     assertFalse(transaction.isSorobanTransaction());
@@ -438,7 +438,7 @@ public class TransactionTest {
             new org.stellar.sdk.operations.Operation[] {operation},
             null,
             new TransactionPreconditions(
-                null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>(), null),
+                null, null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>()),
             null,
             Network.TESTNET);
     assertFalse(transaction.isSorobanTransaction());
@@ -460,7 +460,7 @@ public class TransactionTest {
             new org.stellar.sdk.operations.Operation[] {operation},
             null,
             new TransactionPreconditions(
-                null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>(), null),
+                null, null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>()),
             null,
             Network.TESTNET);
 
@@ -473,7 +473,7 @@ public class TransactionTest {
             new org.stellar.sdk.operations.Operation[] {operation},
             null,
             new TransactionPreconditions(
-                null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>(), null),
+                null, null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>()),
             null,
             Network.TESTNET);
     assertEquals(transaction1.hashCode(), transaction2.hashCode());
@@ -488,7 +488,7 @@ public class TransactionTest {
             new org.stellar.sdk.operations.Operation[] {operation},
             Memo.text("not equal tx"),
             new TransactionPreconditions(
-                null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>(), null),
+                null, null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>()),
             null,
             Network.TESTNET);
     assertNotEquals(transaction1, transaction3);
@@ -502,7 +502,7 @@ public class TransactionTest {
             new org.stellar.sdk.operations.Operation[] {operation},
             null,
             new TransactionPreconditions(
-                null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>(), null),
+                null, null, null, BigInteger.ZERO, 0, new ArrayList<SignerKey>()),
             null,
             Network.PUBLIC);
     assertNotEquals(transaction1, transaction4);